Extract RequestVerificationHeader to its own file.
